Category Update and Delete

+++ admin/categories.php
@@ -24,7 +24,7 @@ include 'closeconnection.php';
             <td class="p-2 border"><?php echo $row['name'];?></td>
             <td class="p-2 border">
                 <a href="editcategory.php?id=<?php echo $row['id']; ?>" class="bg-blue-600 text-white px-2 py-1 rounded-lg mx-1">Edit</a>
-                <a href="actioncategory.php?deleteid=<?php echo $row['id']; ?>" class="bg-red-600 text-white px-2 py-1 rounded-lg mx-1">Delete</a>
+                <a href="actioncategory.php?deleteid=<?php echo $row['id']; ?>" class="bg-red-600 text-white px-2 py-1 rounded-lg mx-1" onclick="return confirm('Are you sure to delete?')">Delete</a>
             </td>
         </tr>
         <?php } ?>
